Add Kodi LED patch
This patch adds the slice LED controls, it is integrated into the GUIWindowHome.cpp
because it can then use a fully integrated Announcement mechanism to trigger the
LED controls.  The other method of doing this would be to use the python events
to provide an LED control but the delay when starting / stopping the video may
be as much as 8 seconds!